Add basic oauth client app
When working with OAuth2 services, especially when developing
first-party implementations, it can be helpful to have a small client
application that can test the full oauth2 flow to ensure protocol
compliance and functionality.

This software should serve this purpose, providing the most minimal
oauth2 client we can to act as a testbed for oauth implementations.